+++ ts/packages/anchor/src/program/namespace/methods.ts
@@ -1,29 +1,31 @@
 import {
-  ConfirmOptions,
   AccountMeta,
+  ConfirmOptions,
+  PublicKey,
   Signer,
   Transaction,
   TransactionInstruction,
   TransactionSignature,
-  PublicKey,
 } from "@solana/web3.js";
-import { SimulateResponse } from "./simulate.js";
-import { TransactionFn } from "./transaction.js";
-import { Idl } from "../../idl.js";
+import { Idl, IdlTypeDef } from "../../idl.js";
+import Provider from "../../provider.js";
 import {
-  AllInstructions,
-  MethodsFn,
-  MakeMethodsNamespace,
-  InstructionAccountAddresses,
-} from "./types.js";
+  AccountsResolver,
+  CustomAccountResolver,
+} from "../accounts-resolver.js";
+import { Accounts } from "../context.js";
+import { AccountNamespace } from "./account.js";
 import { InstructionFn } from "./instruction.js";
 import { RpcFn } from "./rpc.js";
-import { SimulateFn } from "./simulate.js";
+import { SimulateFn, SimulateResponse } from "./simulate.js";
+import { TransactionFn } from "./transaction.js";
+import {
+  AllInstructions,
+  InstructionAccountAddresses,
+  MakeMethodsNamespace,
+  MethodsFn,
+} from "./types.js";
 import { ViewFn } from "./views.js";
-import Provider from "../../provider.js";
-import { AccountNamespace } from "./account.js";
-import { AccountsResolver } from "../accounts-resolver.js";
-import { Accounts } from "../context.js";
 
 export type MethodsNamespace<
   IDL extends Idl = Idl,
@@ -40,7 +42,9 @@ export class MethodsBuilderFactory {
     rpcFn: RpcFn<IDL>,
     simulateFn: SimulateFn<IDL>,
     viewFn: ViewFn<IDL> | undefined,
-    accountNamespace: AccountNamespace<IDL>
+    accountNamespace: AccountNamespace<IDL>,
+    idlTypes: IdlTypeDef[],
+    customResolver?: CustomAccountResolver<IDL>
   ): MethodsFn<IDL, I, MethodsBuilder<IDL, I>> {
     return (...args) =>
       new MethodsBuilder(
@@ -53,7 +57,9 @@ export class MethodsBuilderFactory {
         provider,
         programId,
         idlIx,
-        accountNamespace
+        accountNamespace,
+        idlTypes,
+        customResolver
       );
   }
 }
@@ -78,7 +84,9 @@ export class MethodsBuilder<IDL extends Idl, I extends AllInstructions<IDL>> {
     _provider: Provider,
     _programId: PublicKey,
     _idlIx: AllInstructions<IDL>,
-    _accountNamespace: AccountNamespace<IDL>
+    _accountNamespace: AccountNamespace<IDL>,
+    _idlTypes: IdlTypeDef[],
+    _customResolver?: CustomAccountResolver<IDL>
   ) {
     this._args = _args;
     this._accountsResolver = new AccountsResolver(
@@ -87,7 +95,9 @@ export class MethodsBuilder<IDL extends Idl, I extends AllInstructions<IDL>> {
       _provider,
       _programId,
       _idlIx,
-      _accountNamespace
+      _accountNamespace,
+      _idlTypes,
+      _customResolver
     );
   }